Fix a bug that caused our bot to crash

+++ haliteivbot/rule_based/bot.py
@@ -1304,7 +1304,8 @@ class HaliteBot(object):
                 target = enemy_shipyards[0]
                 # Don't mine
                 if ship.cell.halite > 0:
-                    self.change_position_score(ship, ship.position, -0.8 * self.parameters['move_preference_hunting'])
+                    self.change_position_score(ship, ship.position,
+                                               int(-0.8 * self.parameters['move_preference_hunting']))
                 self.prefer_moves(ship, navigate(ship_position, target.position, self.size),
                                   self.farthest_directions[ship_pos][TO_INDEX[target.position]],
                                   self.parameters['move_preference_hunting'] * 2, penalize_farming,
